backend/requirements.txt
# FastAPI Backend Dependencies
fastapi==0.115.0
uvicorn[standard]==0.30.6
requests==2.32.3
python-multipart==0.0.9
python-dotenv==1.0.0

# Vision Assistant Dependencies (Python 3.13 compatible)
opencv-python>=4.8.1
numpy>=1.26.0
Pillow>=10.1.0
google-generativeai>=0.3.1
